fix "jump to" feature
